backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers>=2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
